And generalised (using index comparison function) unique-sorting and merging
Like |sort@(int,(int,int->bool)| (which has not the reverse-curried form as its
basic definition), this allows |sort_u|, as well as merging auxiliaries for
|sort| and |sort_u|, to be parametrised with a function that compares entries of
a fixed array at given indices, which then can be used to perform these sorting
operations that array, which could be of any type. The sorting proper only
produces a "sorted index" list, which can then be used to shuffle the array.