Improve speed of |rho_stabiliser_orbit|, using "oil spill" method
Like |Levi_normalizer_index|, whose implementation is similar, linear search and
iterative extension of orbits is avoided by producing a stack of ordered orbit
contributions, requiring search only in the top two levels, unique-sorting each
new level after it has been produced. This is still not super fast.

This orbit generation is only necessary it witnesses for W-conjugacy are
desired, which the current code can cater for though the functionality to use it
is not in fact in place yet. For arbitrary Levi subgroups we can explicitly find
a standard Levi subgroup conjugate to it, but for two conjugate standard Levi
subgroups, an explicit conjugation witness can (currently) only be found by
using the |rho_stabiliser_orbit| of one of them.