Add built-ins |rows@mat| |row@(mat,int)|, |root_expression|, and their duals
The |rows| and |row| functions make it more efficient to consider a matrix as a
representation of the lit of its (equal size) rows; for columns this was already
possible, but for readability the column-oriented functions are added as well.
The functions |root_expression| and |coroot_expression| can be useful when
having to deal with the additive structure on the root system.