Make some changes prior to merging into the master branch (and related ones)
This is mainly correcting errors based on previous misunderstanding of the role
of |Symbol| in W_reps_type_BC.at, and some small documentation clarifications.

There are also some Dn related (naming) changes in combinatrics.at

The file W_reps.at requires too much work to synchronise for now, so we just
added a comment that it is outdated; it should be igored in a future merge.