Add built-in functions |is_long_root| and |is_long_coroot|.
Distinguishing long and short roots is a tricky point to implement; although it
is only rarely necessary, it is easy to do it inefficiently or wrongly (indeed a
first version of this commit had a wrong implementation), so we prefer to make
this built-in. Of the two functions |is_long_root| and |is_long_coroot| called
at the same arguments, at most one is true, with both returning false if and
only if the root in question belongs to a simply laced component of the Dynkin
diagram. The algorithm implemented first conjugates the root ot a simple root,
then looks for a witnessing positive coroot for longness, namely whose
evaluation on this root is less than -1; if none is found, |false| is returned.