Remove outdated, unused, and confusing file reductive_character_table.at
This file defines few things, but one of them is |character_table@RootDatum|
that is also defined, and better (because it preserves the given root datum), in
character_tables_reductive.at. It utterly confusing to have the same function
defined differently in similar places, so this script really had to go.