Add function |Levi_A| to combinatorics.at, and use it in ABCD character tables
For a given composition of |n| this function computes |n|, the Lie type of the
corresponding Levi, and the indices of the corresponding simple roots.

The conjugacy class representative generators for simple classical types can use
those indices in order to produce a Weyl word. The function is intended to also
be useful for generating (coweights for) nilpotent orbits in type $A_{n-1}$.