Modified geck_generic.at extensively to run much faster
The main algorithm now maintans a list of the character tables
of all Levis of G, and runs through them inductively, minimizing
dupliation.

For some G the cell version of the computation is faster. Also
there is a trick using \otimes sgn. There are several versions of the
final algorithm, it remains to be seen what is best.