Strip from character_table_reps.at a lot of unnecessary stuff
This involves embedding (which can simply use |convert_to@(RootDatum,WeylElt)|)
and additional |RootDatum| arguments to functions the try to work around a
difference between the desired root datum and the one for which character or
Springer tables are built. Such difference should not exist, and if some code
should be causing such a difference, it should be corrected there, not here.