Fix: revert change that broke |springer_table@RootDatum|; limit |where| output.
A number of unrelated small changes. Notably, as long as |springer_table| uses
|tensor|, have it call |derived| (as before) to avoid producing too large rank
(even though eventually this approach needs to be abandoned). Fix a bug in
|diagram_A_to_partition| (used quotient |q| instead of rest |r|), and another
one in |is_special| (should not test equality of normalised diagram using
(implicitly) matrix equality, as the conversion to matrix is likely to fail for
unequal column lengths). Also modify the |where| diagnostic function to only
print "In call" lines, avoiding clutter and in line with its name;