Remove |n_classes| as member from |CharacterTable|, rename |order| to |order_W|
Since |n_classes| can be obtained from the |class_table| member, it need not be
stored separately. The |order| name could be confused with the size of the
square |table|, so was renamed |order_W|. Finally |n_irreps| is defined as an
alias for |n_classes| what applied to |CharacterTable|, for readability.