Further improve |cross_sub_orbit| by using "bitmap" representation of orbit
This uses assumption that an array of size |KGB_size(G)| does not cause memory
overflow problems (or else probably generating |G| itself would already have).
The "bitmap" is in fact an array of integers where a negative value means an
element is not (yet) in the orbit, and a non-negative value is a sequence number
in generation, and index into the orbit that is under construction. This speeds
up very large cross orbit generation (for instance in high rank complex groups)
but may be sub-optimal for small cross orbits in large KGB sets (because the
sorted orbit is gathered by running over the entire "bitmap" array).