Fix bugs breaking weak_packets: |integral_descent| and |root_permutation|
The function |integral_descent@(SpringerTable,ComplexNilpotent)| ended up
putting a diagram unmodified into a |ComplexNilpotent|; after correcting this
and removing stuff that tries to cure the extinct (or so it should be) disease
of mismatching root systems in Springer tables, it did notinh but call the
function |integral_descent@ComplexNilpotent| from nilpotent_orbits.at, so
easiest was to remove former function and prune calls to it.

The functions |root_permutation| in isomorphism.at not only try to cure the same
disease in a very inefficient way, they also are broken by applying permutations
of size |rank| to the Cartan matrix which is of size |semisimple_rank|. The last
point (only) in fixed here.