Reorganise W_orbit.at somewhat, and use a more systematic naming scheme
Notably, names now reflect whether an initial weight is assumed to be dominant,
and whether the returned value contains more than the main orbit component.

Care has been taken that the dominant-starting-point versions are called (only)
when appreopriate, that generation is mildly efficient for large orbits when it
can, and that duplication of implementation is avoided when it can.