Fix bug in |barycentre_eq|: indexing confusion giving out of bounds access
For the pupose of computing an equation for the barycentric coordinates in an
alcove, the walls are partitioned into diagram components, with a relation
between the coroots for for each component being used to compute the right hand
sides for those coroots. In doing so the index in the relation is that among the
coroots of the component only, but the computed right hand side is placed at the
corresponding position among all wall coroots. The code used the same (wall)
index for both purposes, indexing the relation wrongly and maybe out of bounds.