Remove lusztig_cells.at which is out-of-date int this version; add it back later
This file has such an eventful history that it makes merging branches a pain.
It's history is important though, as the version here removed has been adapted
to changes in the basic structure in a way that the more evolved versions maybe
have not been. Therefore it should be checked that those versions have been
adapted before or when treating the merge conflict (delete/modify) that will
arise when this branch is again merged with versions that do have the file.