Reconstruct lusztig_cells.at from file on commit 636305ef37ec (jeff branch)
This file was reconstructed to ensure that the generic degree computations, that
had been included in the file of this name that was recently removed on the
current branch, were properly taken into account while adapting to the recent
changes. In the end no such computations existed any more, so this is just an
embellished version of the mentioned file on the jeff branch. We did need to
incorporate a new file springer_special_rep.at as well, since this file depends
on it; some other dependency was suppressed by removing two convenience
functions at the end of the file.