Complete typing in w-irrep to special maps for exceptional types: E7 and E8
In spite of the best efforts in Lusztig's book to leave just a bit of ambiguity
in the labelling of the irreducibles for almost every exceptional type, I think
that by closely looking at the mentioned a-values, the proper correspondence can
be found. Currently the date is just in functions like |to_special_E8|; the next
step is to incorporate this into building the character and Springer tables.