Reorganise character tables, simplifying and including a |special| map
The |degrees| member has been replaced by |deg_spec| that also records the
|to_special| map; this member could be made private, since access to these
values is no channelled through |degree| and |special| accessor functions. All
uses of |degrees| in the scripts were immediately subscripted, underlining the
fact that access to a list of all degrees is not really all that useful. The
"special" values are set by calling a function supplied at table construction,
which allows for type-specific functions to be sued for each simple factor;
this is done for all simple types and also in the |combine| function in
character_tables_reductive.at, so that all tables constructed set this field.