Fix and improve the bigint module in utilities
The fix involves |operator*=(int)| that did not properly convert to unsigned for
use of the |mult_add| method. It does require two calls of |negate|, but not on
the same condition: the first must ensure a positive value remains, the latter
must set the correct result sign, taking into account both argument signs.