Improve class template |Free_Abelian_light| for move-only key types
If the key (exponent) type over which the |Free_Abelian_light| container is
built has no copy constructor (and even when it has), one can arrange for
single-term constructors and addition methods to work by moving from their
argument. These methods also move from the coefficient, which is either passed
by value, or is an implied '1' which is then constructed on the spot.