Implement basic functionality for KType, and implicit conversions from it
This allows acces to |x| and |lambda_rho| components, and movement among fibers
to canonical or dominant position, equivalence and normal form for equivalence
and conversion to (Z-linear combination of) final K-types. The implicit
conversions are to Param or ParamPol. While already mentioned in a few places,
the future type KTypePol (with Split coefficients) is not yet defined.