Add more functionality to |Free_Abelian_light|: coefficient assign, |add_term|
We make it possible to directly assign to a given coefficient, or clear the term
of a given key if present (which in contrast to assigning 0 to its coefficient
will not create a term if it was absent, though it will not remove a present
term either). The new form of |add_term| allows moving from the new key.