Change functions in group_operation.at to not generate KGB for their arguments
The functions for product of real forms, and derived and adjoint of a real form
can be defined without generating the KGB structure for the argument real form,
as happens when just about any operation (like |torus_factor| or |involution|
is performed on a KGBElt for that argument. The trick is to use the function
|initial_torus_bits| instead, which just uses the real for (no KGBElt for it)
and allows obtaining a sample torus factor for the argument, which can then be
used to construct a new, synthethic, real form.