Refine somewhat further the interface to KType built-in functions
The new functions that morally accepted or returned |KType| or |KTypePol|
arguments or results now actually do so. Also the |K_Type| defined type in the
scripts has been replaces by |KType|, and in passing usage of |KHighestWeight|
has been strengthened, avoiding exposition of the internals of the definition on
every occasion. Older functions like for branching still do not use |KType|.