Some small script improvements: |gcd@mat|, cross actions by |WeylElt|
This was motivated by editing K_highest_wt.at to better understand the
|project_on_dominant_cone| function there. Now cross actions by a Weyl group
element can be both to the left and right of a |KGBElt| or a |Param| (the
generators to the side of that other argument being applied first) and the
version for |Param| requires an element for the integrality (rather than the
full) datum, since that is the group that acts in this case.

A clerical error breaking |W_orbit@(RootDatum,ratvec)| was alse repaired.