Make W_orbit.at use new built-in orbit function when this is advantageous
This uses weight an coweight versions of |Weyl_orbit|, and for the quotient
groups by a stabiliser (Levi) subgroup, their counterparts |Weyl_orbit_ws|.
In case both the orbit vector and a witnessing Weyl group element are needed,
we use |Weyl_orbit_ws|, since it it easy to get the former from the latter.

Since the built-ins automatically do the necessary for non-dominant arguments,
there are no "|of_dom|" versions for some functions that used to have one; in
that case any clients can simply call the version without the "|of_dom|".