Added facet_vertsMVL@(RootDatum,ratvec), sphericalMVL@Param
First of these uses Marc's function wall_set@(RootDatum,ratvec) to
compute an alcove in which the ratvec lies, and then linear algebra to
compute the vertex set of the corresponding facet. This is now about
50% slower than the old code (which always used the affine Weyl group
to translate back to the fundamental alcove). If my script were
cleaned up, then perhaps the speed disadvantage would decrease; or if
I had access to the library function wall_set, then CERTAINLY the
speed disadvantage would decrease.