small organization and speedup
Moved most unused functions to the end of the file; main content is
now up to definition of spherical. Added more explanation of math in
intro. Significant speedup (factor more than 2 in rank 4) of
FPP_vlist_indices@RootDatum, by replacing a search for an affine W
element by a search for a finite W element.

Suspect that testing many Params for is_unitary will be sped up by
doing _more_ alcove_center in deformation. This is a bit painful to
test.