Remove |l_shifts| from |Ext_common_context|, and |ext_gen| from |complex_cross|
The goal here is to ultimately remove the integral-subsystem dependency from
|Ext_common_context| and from the |star| function: we prefer the latter uses a
|RootNbr| for the full root datum, with a length indication (1,2, or 3), instead
of an |ext_gen|, so as to make it agnostic of whether the caller thinks in terms
of the integrality datum (as when constructing an extended block) or not (as
when doing things like |scaled_extended_dominant|). Notably we think of a
variant of |scaled_extended_dominant| where we know the full system is integral,
and want to avoid the awkwardness of computing an "integral system" for it.

In passing another old camelCase was swatted: |reflectionWord| no longer is.