Reorganise ext_block so that |Ext_common_context| no longer stores a |SubSystem|
This has as a consequence that use of |ext_gen| storing simple indices into the
(integral) subsystem can no longer be used at the interface to methods that used
the |SubSystem| from their |Ext_common_context| to interpret the |ext_gen|
values. Instead, they now take a |RootNbr| value, index into the full root
datum, for one of the roots of the |ext_gen| (the other can be found using
|delta_of| if necessary), plus possibly a length (1,2, or 3) so as to avoid the
need to recompute this value repeatedly. Some other small changes to
|Ext_common_context| we also needed to make the program compile correctly.