Add a function for converting K-type polynomials from tabled to external form
For memory efficiency reasons, a |Rep_table| stores each needed |K_type| only
once, and therefore has a special format for K-type polynomials that only stores
an index to a K-type for each term. In the interpreter, like for the branching
code, K-type polynomials contain actual K-types that are sorted, principally by
height. A conversion function |export_K_type_pol| that quickly gathers and then
sorts the terms is defined, and used in |full_deform| and |twisted_full_deform|.