Fix wrong change to |twisted_full_deform| accidentally made in previous commit
The wrapper function used to correctly call the |extended_finalise| method to
expand its argument into final parameters (so taking into account flips to the
default exptension that occur in the process) but by sloppy editing this had
been changed to call the |finals_for| method instead. This change is reverted.