Replace old |Rep_context::finals_for(StandardRepr)| by new implementation
All remaining uses of the old function were replaced by calls to the new one
(which can incidentally handle non-standard arguments, so that |deform| and
|full_deform| no longer insist that their argument be standard), which returns a
list of parameters with a coefficient (for a possible sign), so that callers now
take the latter into account. The new method that used to be called |finals|,
was renamed |finals_for| after removal of the old method of that name.