Rewrite |extended_finalise| method to only require parameter to be standard
This means that the method must cope with non-dominance for complex and real
simple roots (the wrapper function did not test for this, which could lead to
assertion failure and crash), in addition to singular descents of all kinds
(which it already did). The implementation has been fundamentally changed,
as one cannot assume the infinitesimal character dominant and unchanging; on the
other hand there is no need to construct the singular subsystem, and all
reflections considered are simple ones. Dealing with non-dominance for real
simple roots was not previously done at all in the extended setting, so this
required using new untested formulas, which for the time being are experimental.