Add function |scaled_extended_finalise| and use it for twisted deformation
This function replaces a combination of |scaled_extended| and
|extended_finalise|, and uses a simplified mechanism that just repeatedly test
complex simple roots for either non-dominance of |gamma|, or for singular
(complex) descents. The result is always a single parameter, and maybe a flip.

This change makes |scaled_extended| and |extended_finalise| all but redundant;
they are still directly callable as separate built-ins though. Ultimately the
the built-in interface should use the same methods as are used internally.