Add Boolean argument to |Free_Abelian_light| constructor from |std::vector|
When constructing internally a polynomial from a vector of terms, there may or
may not be need to sort the terms; in general this sorting is needed to preserve
the class invariant, but in some cases it is a waste of time, for instance when
forming a polynomial from a range of terms in an existing one. We now provide a
Boolean argument telling whether sorting is requested. The old situation was to
always sort; currently no default is provided for the Boolean argument, but if
one is added in the future, it should obviously be |true| (when in doubt, sort).