Fix inconsistency in ext_kl unit: its |KLIndex| should be |ext_kl::KLIndex|
In Atlas.h we define two separate type definitions for |KLIndex|, one in the
namespace |kl| and another in the namespace |ext_kl|. This allows using
different size of (unsigned) integers in the hash tables for storing KL
polynomials in the ordinary and extended context (the latter presumably is less
likely to require as many distinct polynomials as the former). However, for this
to work, the ext_kl compilation unit should systematically use |ext_kl::KLIndex|
but it did not. In fact on most occasions it explicitly referenced |kl::KLIndex|
with no explanation (simply using |KLIndex| would have selected the definition
from its own namespace). (It would be an interesting historic research to find
out how these explicitly wrong references came into existence originally.) We
have now changed these indications to |ext_kl::KLIndex|, so as to be quite clear
that these are what we must use: without these changes, one cannot compile the
ext_kl compilation unit correctly unless Atlas.h defines the two types to name
the same integer type (as is does currently, but which restriction of course
defeats the purpose of having separate definitions in the first place).

Ultimately one might prefer to simply use |KLIndex| in the affected instances.

LocalWords:  KLIndex