Reorganise identifier completion and provide |readline_completions| as built-in.
The reorgansation was necessary since, due to the weird interface of |readline|,
so separate function was present to just generate a list of completions for a
prefix (instead a function needs to provide a different completion on each
successive call). After the canges a separate new function serves both the
actual completion for |readline| and the implementation of a new built-in
function |readline_completions| that caters mostly for non-human customers.