Added a function "theta_stable_quasi_data@Param p"
This returns (Q,pL) with Q=LU theta-stable, L quasisplit, so that
cohomological induction of pL from Q gives the original p on
G. This Q is UNIQUE/canonically associated to p, unlike
the split version given by theta_stable_data. Consequence is that the
LKTs of pL and of p are in bijection.

This required an additional function from_dlambda_dom@(Param) in
Weylgroup.at. No changes to any existing functions.