Use new bitset functionality in user functions defined in basic.at
There is some modified functionality: the function |contains@int| now produces a
predicate on |vec| rather than |int| values, and they should not have negative
entries. The functions |is_subset_of_sorted| and |is_disjoint_from_sorted| are
retained, but new functions |is_subset_of| and |is_disjoint_from| with vector
arguments (that are converted to bitsets, so need not be sorted) are to be
preferred especially for large (finite) sets of natural numbers. There are
uncurried versions |is_subset| and |is_disjoint| between pairs of vectors, and a
new version |contains@vec| producing a superset predicate. A new function
|set_bit_positions| provides a section (right-inverse map) of |to_bitset|.