Define optimised search functions |index_of| for lists of |vec| or |ratvec|.
Supposing all (rational) vectors have the same size, these function are faster
than linear search if multiple searches into the same list have to be performed.
For this reasong these functions are necessarily curried, i.e., they take just
the list as argument and produce a function to which individual elements can be
(repeatedly) be presented as argument for looking them up.

It is achieved by constructing a |KTypePol| or |ParamPol| from the list, where
searching a term can be done in logarithmic time. This (ab)uses |RealForm|
values for the compact torus (for |vec|) or the split torus (for |ratvec|) as
basis for building polynomials (of K-types respectively parameters). A similar
(but more straightforward) function for searching a list of |Param| values (all
for the same real form) was already present under the name |find_in|; it has
been renamed |index_of| so that the name |find| always indicates linear search.