Start reorganising facetious.at for more systematic use of data types
Notably a vertex list and its lookup function have been grouped together
in a |VertexData| structure inside a |FacetData|, the latter also defining the
facets of interest (by their vertex indices) and their status. Many function
interfaces have been modified to use |VertexData| instead of a raw |[ratvec]|;
for now the extra search functionality is not often used, but the change should
start to improve the transparency of what is being done.