Fix bound checks for indexing and slicing, for large index values.
Bound checks in indexing expressions |a[i]| and slicing |a[i:j]| were performed
after getting |i| (and |j|) through |int_val| and after converting the size of
|a| to (signed) |int| as well. This gave wrong results and strange error
messages when the size, and possibly index, pass the 2^31 mark. Instead we now
use (the 64-bit) |long_val| instead of |int_val|, and properly separated signed
and unsigned tests to detect and report errors, allowing indices up to 2^63.