Added Annegret's anti_finalize to goodroots.at
Used antifinalize in a version of is_pos_on_LKTsAP. It tests almost
exactly the same speed as is_pos_on_LKTs; probably easy to clean up
the code?

Tweaks of facetious. updateSing now includes all larger facets on
which the deformation is constant (previously it was those where it's
"obviously" constant). updateSingClosure now records also smaller
facets in the closures of deformed larger facets that are being added.