added induction with chopped facets
New function spherical_regFold(G) uses induction from maximal Levis to
account for lots of unitary reps. It's still kind of a mess, notably
when cofolding is involved; but I think it may give correct unitarity
answers even when it fails to deal with induction properly.

Small edits to chopping_facets to allow it to be included in facetious.