Speed spherical_reg using chopping_facets_fast
Also added keeping tracking of restrictions to K of all irreducibles
being studied (in field Kchars = [[KTypePol]] of FacetDataKchar, indexed in
parallel with facet_status_inds [[FacetStatusVertsIndex]]). I think
this is not helpful for speed, because the computation of Kchars is
mostly stored internally, so not repeated. But it's meant to help
think about equivalence of facets.