New script WeylTest.at sketches the difficulties in FPPopenureAlcoves
In big G, lots of time in spherical is spent in the function
FPPopenureAlcoves, which takes a facet Base in the FPP and returns all
the alcoves in the FPP containing Base in their closure. The current
code just runs over all the alcoves looking for winners. I'm trying
write something that _constructs_ the correct alcoves, in the hope
that it could be faster. The function openureAlcoves@(RootDatum,
ratvec) in WeylTest.at sketches two constructions. Both appear to be
slower than the current FPPopenureAlcoves; but I'm hopeful...