Add functions to facet_classes.at to reduce facet set by obvious equivalence
This equivalence is for connected components of the undirected graph with edges
for facet-face pairs whose K-characters are equal. This appears to be at least
very close, if not equal, to the equivalence that |strong_components| will find
for the directed graph where in addition to these bidirectional edges the
directed face inclusions are added, and connected component discovery is quite
easy. Potentially inserting |to_equivalence| before calling |strong_components|
can be faster than computing the strong components directly by Tarjan.