Modify |add_links| in graph.cpp to use |std::set| instead of |BitMap| data type
This is one place where the implementation of Tarjan's algorithm can get
unnecessarily slow for large graphs: to unite links that point into the same
component to a single link, the (Atlas) |BitMap| data type was used to record
the components reachable from a given component in the induced graph. But this
requires a bitmap whose size is the (current) number of strong components, and
this increases linearly with the size of the induced graph. To get a cost that
is (only) proportional to the actual number of links to be processed (times the
logarithm of the number of reachable components), we now use the |std::set|
(binary tree) data type instead. This could speed up for very large graphs.