Change |RowBase| in |Transducer| to a |std::array| instance; easier and cleaner
The only thing that the defined type was addint with respect to |std::array| was
default initialisation to explicit Undef values. The implementation was changed
to explicitly fill with such values where is was necessary; this was unavoidable
since |std::array| does not allow it inside its constructor, but also clearer.