Reorganise Weyl group module step 2: use less data in each |Transducer|
Notably we no longer store a full |WeylWord| (vector) for each piece (state)
in the transducer, just the last letter (unless for the zero piece). From this
the word can be deduced by repeatedly shifting down producing a generator to the
right, until the piece becomes zero. Often, when generators are needed in order
from right to left, this can be applied directly; when generators are needed
from left to right, we do a rapid reversal using a stack. To make this fit in a
fixed-size local array, we define a constant for the largest possible length of
a single piece (independently of the type of Weyl group). In order to make the
computation of this maximum |constexpr| we must advance to the C++14 version.