Fix bugs in three extended scale/finalise functions: must loop over orbits
The code used a loop up to |rd.semisimple_rank()|, so over the nodes of the
(unfolded) Dynkin diagram, but they should loop over |delta|-orbits of those
nodes (so over the folded diagram). The error was obvious in the fact that the
loop variable was then used as an index for orbits. It remains somewhat
mysterious that such an obvious error should not have caused crashes long ago,
as in any unequal rank case, running the loop to completion would seem to
produce an out-of-bounds access of the |orbits| vector. Although often the loop
does get interrupted, it is hard to imagine that it always did, even if the
Dynkin diagram was always connected (the bug was found in a disconnected case).