Make |Transducer| table only depend on its Dynkin diagram component and place
In other words, the |Transducer| tables for a component of the Dynkin diagram
depend only on the type of that component, not on the presence of other
components to its left (generators for such components used to be passed through
the transducer unchanged, but now they are never even seen by the transducer,
and its tables are correspondingly smaller). It does require yet another
numbering scheme for generators, as the local indexing scheme within a
componenent is lower by an offset than the indexes of transducers thenselves and
of |WeylElt| pieces. At certain points a change of interpretation happens, and
there an appropriate shift is applied to the index values.