Replace |shift| and |out| fields in |Transducer| by constructor-local vector
Now the fields of the per-piece structure for shift and output actions are
removed once the matrix is constructed, and these actions on a completed
transducer use the matrix. This does mean that during construction these action
methods cannot be called, and direct access to a local vector |tab| that is
ultimately turned into the action table replaces it. The auxiliary methods for
traversal of the dihedral structures during construction have been expanded into
the constructor and then removed, as it would be too cumbersome to get access to
the constructor-local vector to implement them as method or function.