Remove Coxeter matrix member from |WeylGroup|, decide commutation from table
Since the diagram is stratified in a standard manner, it is possible to compute
the Coxeter matrix entries needed while constructing the transducers from the
type (determined in the Dynkin module) and the component-local indices directly,
which a new function |Coxeter_entry| does. The only other purpose for which the
Coxeter matrix was used is to determine which generators commute while
computing the |d_min_start| values; with some thought, this can be seen to be
determined by whether a specific transduction step involved in multiplying the
larger |Generator| (as |WeylElt|) from the right by the smaller generator
involves as shift (if so there is no commutation). This is now used near the end
of the |WeylGroup| constructor, and also to implement the public method
|commutes| (for which in addition conversion to inner numbering is done first).