Implement built-ins |basic_orbit_ws|, |affine_orbit_ws| for orbits modulo roots
At an internal level two kinds of subquotients can be generated: for a pair of
contained pseudo-Levi subgroups of successive ranks (the fundamental work is
done in |basic_orbit|, using only the Cartan matrix of the larger pseudo-Levi),
and for a pair of a full rank pseudo-Levi and the full Weyl group (here the
work is done in |vertex_orbit|, so called because it traces the orbit of a
nonzero vertex of the fundamental alcove, using the Cartan matrix and a label).

At an external level this is packaged into the two mentioned built-ins, where
|basic_orbit_ws| generates a single-step subquotient, while |affine_orbit_ws|
combines several successive subquotients to a quotient of the full Weyl group
by a pseudo-Levi that is the integral Weyl group of a given rational vector.
In both cases a list of |WeylElt| values is returned, leaving their application
to (rational) weights up to the user. The internal function return lists of
|orbit_elem| structures that can be converted to lists of coset representatives.

This commit also changes the built-in |walls|, reordering the returned root
indices such that the integral ones come first (with an extra result component
telling how many there are) and listing the remaining roots by decreasing labels
for the alcove, which is the ordering internally used in |affine_orbit_ws|.