Make changes in matrix and bitmap utilities, avoids uses of bare |unsigned int|
For |BitMap| the numbers "stored" (used internally as indices of bits) were of
type |unsigned long|, which might be only 32 bits, which has been replaced by
the type |size_t|; there are some similar changes for |big_int| as well. For
matrices no change to the the 32-bit row/column indices is made yet, as making
them 64 bits would increase storage requirements by 4 bytes per matrix, but a
defined type |index_t| was used to identify the places where |unsigned int|
was used as index type, so that it becomes easier to change in the future, if
extremely deep or wide matrices might be called for.

In passing, the comparison between |big_int| and machine integer, presumably
most used with a small constant in the second position, was changed from having
an unsigned to a signed integer as second argument, as this seems more natural
and allows testing for -1 without fear (although in the old situation conversion
of the argument to |unsigned| would have done the proper test as well). The
|is_one| method was replaced by usage of this |==|, but |is_zero| was kept.

Finally we took the occasion to unCamelCase |numRows| and |numColumns|.