Change global.w: ensure large integer arguments are not truncated inadvertently
There are not much occasions here where large integer arguments would be very
useful in practice, so often it is a matter of providing more consistent limits
with clear error message; for instance all value used for indexing matrices must
currently fit in 32 bits, but they are first converted to unsigned 64 bits
(which will refuse negative values with a conversion error) an then signal an
implementation limit if they are too large for their use with matrices.

One real change is allowing |null| to create a vector of $2^32$ entries or more,
which used to be impossible. Error testing is necessary however each time a
vector size is transformed into a matrix dimension, for instance when
transposing the vector to a single-row matrix.

Also the implementation of (integer) powers of integers or rationals was
somewhat streamlined, handling all cases that don't actually overflow correctly.