Use MVL new W-orbit generation in facetious.
I thought that I had committed these changes a few days ago, and I
threw away the files where the changes were made. I recovered this
facetious.at from a backup; it SEEMS to be the right thing, but there
is some danger that I forgot about an important change to another file
or something like that.