Fix an old bug (DV reported Sep 11 2021) breaking root data with >256 posroots
The method |RootSystem::extend_to_roots| called |simple_reflected_root| to get
the root permutation for reflection for the _image_ of a simple root (by some
root system automorphism), after the root number of that image was converted by
|RootSystem::rt_abs| to a positive root index. Because |simple_reflected_root|
internally indexes an array that is has entries for any positive root index,
this worked, even though the image is generally not a simple root, as long as
the positive root index would fit in the |weyl::Generator| type, that has 8
bits. For large root data that fails, and the wrong root permutation was used.
The correction is to use |reflected_root| instead; this simplifies the code.