Fix some remaining errors with counted integer loops needing more than 32 bits
There was still one internal loop counted devlared as |int|. While we were at it
we made the system robust against loops where the counter overflows the 64-bit
signed integer boundary, signalling an error at the start-up of the loop it the
specified range would provoke such an error. This does cost an extra cycle
(outside the lopp body) for every exucution of a counted loop, which is too bad.