minor SL(odd,R) corrections, add sphericalKPO_big to facetious
There was a very old bug in the case of a folded A_even diagram, which
returned an FPP with some extra facets (not really in the FPP). These
facets were of course found to be nonunitary, so no harm done.

The _big functions sphericalKPO_big and spherical_testKPO_big keep
track of root lattice translates bup, bdown for each FPP facet, so
that (when computing K-characters) KLV computations always use an
integrality datum for a facet of the fundamental alcove; so the NUMBER
of integrality data that must be stored is bounded by 2^{rank +
1}. This can provide substantial time and memory savings for
K-characters of low-dimensional facets. These shifts are NOT used in
the is_unitary calculations, so the memory savings is mostly
eliminated during those.